refactor(core): deprecate WithReferringDocuments HOC in favor of a useReferringDocuments() hook
