+++ packages/sanity/package.json
@@ -208,7 +208,6 @@
     "react-fast-compare": "^3.2.0",
     "react-focus-lock": "^2.8.1",
     "react-is": "^18.2.0",
-    "react-props-stream": "^1.0.0",
     "react-refractor": "^2.1.6",
     "react-rx": "^2.1.3",
     "read-pkg-up": "^7.0.1",
@@ -217,7 +216,7 @@
     "rimraf": "^3.0.2",
     "rxjs": "^7.8.0",
     "rxjs-etc": "^10.6.0",
-    "rxjs-exhaustmap-with-trailing": "^1.2.0",
+    "rxjs-exhaustmap-with-trailing": "^2.1.0",
     "sanity-diff-patch": "^1.0.9",
     "scroll-into-view-if-needed": "^3.0.3",
     "semver": "^7.3.5",
