+++ packages/sanity/src/core/components/WithReferringDocuments.tsx
@@ -1,48 +1,22 @@
 import {SanityDocument} from '@sanity/types'
-import {withPropsStream} from 'react-props-stream'
-import {concat, Observable, of} from 'rxjs'
-import {distinctUntilChanged, map, switchMap} from 'rxjs/operators'
+import {ReactElement} from 'react'
 import {DocumentStore} from '../store'
+import {useReferringDocuments} from '../hooks/useReferringDocuments'
 
-/** @internal */
-export const WithReferringDocuments = withPropsStream(
-  connect,
-  function _WithReferringDocuments({children, ...props}) {
-    return children(props)
-  }
-)
-
-function connect(
-  receivedProps$: Observable<{
-    children: (props: {
-      documentStore: DocumentStore
-      id: string
-      isLoading: boolean
-      referringDocuments: SanityDocument[]
-    }) => React.ReactElement
-    documentStore: DocumentStore
-    id: string
-  }>
-) {
-  return receivedProps$.pipe(
-    distinctUntilChanged((prev, next) => prev.id === next.id),
-    switchMap((receivedProps) =>
-      concat(
-        of({...receivedProps, referringDocuments: [], isLoading: true}),
-        receivedProps.documentStore
-          .listenQuery(
-            '*[references($docId)] [0...101]',
-            {docId: receivedProps.id},
-            {tag: 'with-referring-documents'}
-          )
-          .pipe(
-            map((docs: any[]) => ({
-              ...receivedProps,
-              referringDocuments: docs,
-              isLoading: false,
-            }))
-          )
-      )
-    )
-  )
+/**
+ * @internal
+ * @deprecated - Will be removed in 4.0.0, use the `useReferringDocuments(<documentId>)` hook instead
+ */
+export function WithReferringDocuments({
+  children,
+  id,
+}: {
+  children: (props: {isLoading: boolean; referringDocuments: SanityDocument[]}) => ReactElement
+  /**
+   * @deprecated - no longer required
+   */
+  documentStore?: DocumentStore
+  id: string
+}) {
+  return children(useReferringDocuments(id))
 }
